+++ Sources/SkipBridge/Skip/Reflector.kt
@@ -4,7 +4,7 @@
 // under the terms of the GNU Lesser General Public License 3.0
 // as published by the Free Software Foundation https://fsf.org
 
-package tools.skip.bridge.kt
+package tools.skip.bridge
 
 import kotlin.reflect.*
 import kotlin.reflect.full.*